fix(cli): force the tui when --output == tui (#331)
# Summary

Fixes an issue that prevents the TUI from showing when the user config
disables interactive mode but the output flag specifies `tui`. With this
change, the output format flag should always win over the config value.
The `DISABLE_FLOW_INTERACTIVE` env var is still the used to disable the
TUI with that flag set.

**Notable Changes**

- Removes `tui` as the default output format
- Respects the value of `--output=tui` when the user configs disables
the TUI

Closes #330